fix: rename file to camelCase
