+++ src/components/AccesoryCardList.jsx
@@ -1,4 +1,4 @@
-import { AccesoriesData } from "../constants/AccesoriesData";
+import { AccesoriesData } from "#/constants/accesoriesData";
 import Image from "next/image";
 
 export function AccesoryCardList() {
