+++ src/components/sections/AskQuestions.section.jsx
@@ -4,7 +4,7 @@ import { SectionTitle } from "../SectionTitle";
 import Link from "next/link";
 import * as Solid from "@heroicons/react/24/solid";
 import * as Out from "@heroicons/react/24/outline";
-import { AnchorSectionNames } from "#/constants/AnchorSectionNames";
+import { AnchorSectionNames } from "#/constants/anchorSectionNames";
 
 export function AskQuestionsSection() {
 	return (
